make dwm installable
